Now detects in-progress recordings, and mark them in the user interface.
Also sets the duration based on the scheduled end, not the actual end-of-recording,
which should allow users to do "timeshift playback", i.e. start watching
a recording just after it started

However, MythTV still seems to stop the playback stream at the point
where it ended when the playback started. But now this code does its part...